reworked the way colors are picked to make sure they contrast well
it actually turned out somewhat nice! You have a 50% chance of getting a dark
theme and a 50% chance of getting a light theme. It might be nice to have some
kind of switch or config file for if you favor one over the other, but overall
things turned out pretty well!

The main thing I'm concerned about now is that there are more colors that modern
terminals have set than just foreground and background. I would be surprised if there
weren't potential clashes with those that make certain colored terminals unusable.

Generating certain color schemes shouldn't be that hard, though, in order to get around
that. It might be nice to use some kind of prng instead of this hacky crc32 trick in
order to get seeding datat for these colors...